Move Nuclide warning before the transpiler hook
